Webpack

+++ src/controllers/videoController.js
@@ -13,7 +13,6 @@ export const home = async (req, res) => {
 export const watch = async (req, res) => {
     const { id } = req.params;
     const video = await (await Video.findById(id)).populate("owner");
-    console.log(video);
     if(!video)
         return res.render("404", {pageTitle: "Video Not Found"});  
     return res.render("watch", {pageTitle: `Watching ${video.title}`, video}); 
@@ -24,6 +23,9 @@ export const getEdit = async (req, res) => {
     const video = await Video.findById(id);
     if(!video)
         return res.status(404).render("404", {pageTitle: "Video Not Found"});  
+    if(String(video.owner) !== String(req.session.user._id)){
+        return res.status(403).redirect("/");
+    }
     return res.render("edit", {pageTitle: `Editing${video.title}`, video})
 };
 export const postEdit = async (req, res) => {
@@ -32,6 +34,9 @@ export const postEdit = async (req, res) => {
     const video = await Video.exists({_id:id});
     if(!video)
         return res.status(404).render("404", {pageTitle: "Video Not Found"});  
+    if(String(video.owner) !== String(req.session.user._id)){
+        return res.status(403).redirect("/");
+     }
     await Video.findByIdAndUpdate(id, { // await 안쓰니까 작동을 안했음 꼭 쓰자
         title: title,
         description: description,
@@ -60,13 +65,16 @@ export const postUpload = async (req, res) =>{
     const file = req.file;
     const { title, description, hashtags } = req.body;
     try{
-    await Video.create({
+    const newVideo = await Video.create({
         title,
         description,
         owner:_id,
         fileUrl:file.path, 
         hashtags: Video.formatHashTags(hashtags)
     });
+    const user = await User.findById(_id)
+    user.videos.push(newVideo._id);
+    user.save();
     return res.redirect("/");
     } catch(error){
         return res.status(400).render("Upload", {pageTitle: "Upload Video", errorMessage: error.message} );
@@ -74,6 +82,12 @@ export const postUpload = async (req, res) =>{
 }
 export const deleteVideo = async (req, res) =>{
     const { id } = req.params;
+    const video = await Video.findById(id);
+    if(!video)
+        return res.status(404).render("404", {pageTitle: "Video Not Found"});  
+    if(String(video.owner) !== String(req.session.user._id)){
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect("/")
 }
